+++ src/EdFi.AnalyticsMiddleTier.DataStandard31/Asmt/MSSQL/0013-View-StudentAssessmentObjectiveFact-Drop.sql
@@ -1,0 +1,10 @@
+-- SPDX-License-Identifier: Apache-2.0
+-- Licensed to the Ed-Fi Alliance under one or more agreements.
+-- The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
+-- See the LICENSE and NOTICES files in the project root for more information.
+
+IF EXISTS (SELECT 1 FROM INFORMATION_SCHEMA.VIEWS WHERE TABLE_SCHEMA = 'analytics' AND TABLE_NAME = 'asmt_StudentAssessmentObjectiveFact')
+BEGIN
+	DROP VIEW analytics.asmt_StudentAssessmentObjectiveFact
+END
+GO
