+++ src/EdFi.AnalyticsMiddleTier.DataStandard32/Asmt/PostgreSQL/0012-View-StudentAssessmentFact-Drop.sql
@@ -1,0 +1,6 @@
+﻿-- SPDX-License-Identifier: Apache-2.0
+-- Licensed to the Ed-Fi Alliance under one or more agreements.
+-- The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
+-- See the LICENSE and NOTICES files in the project root for more information.
+
+DROP VIEW IF EXISTS analytics.asmt_StudentAssessmentFact;
