+++ src/EdFi.AnalyticsMiddleTier.Tests/TestCases/StudentAssessmentFact/MP_2013_Mathematics_Seventh_Grade_605213_255901044_should_have_AssessmentKey.xml
@@ -4,8 +4,8 @@
   <ControlDataInsertion>
   </ControlDataInsertion>
   <Query>
-      SELECT ReportingMethod
-      FROM analytics.asmt_StudentAssessmentObjectiveFact
+      SELECT TOP 1 AssessmentKey
+      FROM analytics.asmt_StudentAssessmentFact
       WHERE StudentObjectiveAssessmentKey='392-Seventh grade Mathematics-48-MP-2013-Mathematics-Seventh grade-zXFMqaQUouumDhHF8VnSmBJPaZ1Xbc8f/Li1nh1q-uri://ed-fi.org/Assessment/Assessment.xml'
       AND StudentSchoolKey='605213-255901044'
       AND SchoolKey='255901044'
@@ -14,6 +14,6 @@
       AND ReportingMethod='Raw score';
   </Query>
   <Result>
-      <ReportingMethod>Raw score</ReportingMethod>
+      <AssessmentKey>MP-2013-Mathematics-Seventh grade-uri://ed-fi.org/Assessment/Assessment.xml</AssessmentKey>
   </Result>
 </TestCase>
