+++ src/EdFi.AnalyticsMiddleTier.Tests/TestCases/StudentAssessmentFact/MP_2013_Mathematics_Seventh_Grade_605213_255901044_should_have_ReportingMethod.xml
@@ -4,8 +4,8 @@
   <ControlDataInsertion>
   </ControlDataInsertion>
   <Query>
-      SELECT COUNT(1) as CountValue
-      FROM analytics.asmt_StudentAssessmentObjectiveFact
+      SELECT TOP 1 ReportingMethod
+      FROM analytics.asmt_StudentAssessmentFact
       WHERE StudentObjectiveAssessmentKey='392-Seventh grade Mathematics-48-MP-2013-Mathematics-Seventh grade-zXFMqaQUouumDhHF8VnSmBJPaZ1Xbc8f/Li1nh1q-uri://ed-fi.org/Assessment/Assessment.xml'
       AND StudentSchoolKey='605213-255901044'
       AND SchoolKey='255901044'
@@ -14,6 +14,6 @@
       AND ReportingMethod='Raw score';
   </Query>
   <Result>
-      <CountValue>1</CountValue>
+      <ReportingMethod>Raw score</ReportingMethod>
   </Result>
 </TestCase>
