+++ src/EdFi.AnalyticsMiddleTier.Tests/TestCases/StudentAssessmentFact/MP_2013_Mathematics_Seventh_Grade_605213_255901044_should_have_StudentScore.xml
@@ -4,14 +4,16 @@
   <ControlDataInsertion>
   </ControlDataInsertion>
   <Query>
-      SELECT COUNT(1) as CountValue
-      FROM analytics.asmt_StudentAssessmentObjectiveFact
+      SELECT TOP 1 StudentScore
+      FROM analytics.asmt_StudentAssessmentFact
       WHERE StudentObjectiveAssessmentKey='392-Seventh grade Mathematics-48-MP-2013-Mathematics-Seventh grade-zXFMqaQUouumDhHF8VnSmBJPaZ1Xbc8f/Li1nh1q-uri://ed-fi.org/Assessment/Assessment.xml'
-      AND StudentSchoolKey='605213-255901001'
+      AND StudentSchoolKey='605213-255901044'
+      AND SchoolKey='255901044'
+      AND StudentScore='1'
       AND ResultDataType='Decimal'
-      AND ReportingMethod='Achievement/proficiency level';
+      AND ReportingMethod='Raw score';
   </Query>
   <Result>
-      <CountValue>0</CountValue>
+      <StudentScore>1</StudentScore>
   </Result>
 </TestCase>
