+++ src/EdFi.AnalyticsMiddleTier.Tests/TestCases/StudentAssessmentFact/MP_2013_Mathematics_Seventh_Grade_605213_255901044_should_return_one_record.xml
@@ -1,0 +1,20 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<TestCase>
+  <DBMS>Any</DBMS>
+  <ControlDataInsertion>
+  </ControlDataInsertion>
+  <Query>
+    SELECT COUNT(1) as CountValue
+    FROM analytics.asmt_StudentAssessmentFact
+    WHERE StudentAssessmentKey = 'MP-2013-Mathematics-Seventh grade-uri://ed-fi.org/Assessment/Assessment.xml-zXFMqaQUouumDhHF8VnSmBJPaZ1Xbc8f/Li1nh1q-392'
+    AND StudentObjectiveAssessmentKey='392-Seventh grade Mathematics-48-MP-2013-Mathematics-Seventh grade-zXFMqaQUouumDhHF8VnSmBJPaZ1Xbc8f/Li1nh1q-uri://ed-fi.org/Assessment/Assessment.xml'
+    AND StudentSchoolKey='605213-255901044'
+    AND SchoolKey='255901044'
+    AND StudentScore='1'
+    AND ResultDataType='Decimal'
+    AND ReportingMethod='Raw score';
+  </Query>
+  <Result>
+      <CountValue>1</CountValue>
+  </Result>
+</TestCase>
