+++ src/EdFi.AnalyticsMiddleTier.Tests/TestCases/StudentAssessmentFact/MP_2013_Mathematics_Seventh_Grade_605215_255901044_should_have_PerformanceResult_Basic.xml
@@ -4,14 +4,14 @@
   <ControlDataInsertion>
   </ControlDataInsertion>
   <Query>
-      SELECT COUNT(1) as CountValue
-      FROM analytics.asmt_StudentAssessmentObjectiveFact
+      SELECT PerformanceResult
+      FROM analytics.asmt_StudentAssessmentFact
       WHERE StudentObjectiveAssessmentKey='396-Seventh grade Mathematics-48-MP-2013-Mathematics-Seventh grade-zXFMqaQUouumDhHF8VnSmBJPaZ1Xbc8f/Li1nh1q-uri://ed-fi.org/Assessment/Assessment.xml'
       AND StudentSchoolKey='605215-255901044'
       AND ResultDataType='Decimal'
       AND ReportingMethod='Raw score';
   </Query>
   <Result>
-      <CountValue>1</CountValue>
+      <PerformanceResult>Basic</PerformanceResult>
   </Result>
 </TestCase>
