+++ src/EdFi.AnalyticsMiddleTier.Tests/TestCases/StudentAssessmentFact/MP_2013_Mathematics_Seventh_Grade_605215_255901044_should_return_one_record.xml
@@ -4,14 +4,14 @@
   <ControlDataInsertion>
   </ControlDataInsertion>
   <Query>
-      SELECT PerformanceResult
-      FROM analytics.asmt_StudentAssessmentObjectiveFact
+      SELECT COUNT(1) as CountValue
+      FROM analytics.asmt_StudentAssessmentFact
       WHERE StudentObjectiveAssessmentKey='396-Seventh grade Mathematics-48-MP-2013-Mathematics-Seventh grade-zXFMqaQUouumDhHF8VnSmBJPaZ1Xbc8f/Li1nh1q-uri://ed-fi.org/Assessment/Assessment.xml'
       AND StudentSchoolKey='605215-255901044'
       AND ResultDataType='Decimal'
       AND ReportingMethod='Raw score';
   </Query>
   <Result>
-      <PerformanceResult>Basic</PerformanceResult>
+      <CountValue>1</CountValue>
   </Result>
 </TestCase>
