--- src/EdFi.AnalyticsMiddleTier.Tests/TestCases/StudentAssessmentObjectiveFact/0001_StudentAssessmentObjectiveFact_should_match_column_dictionary.xml
@@ -1,54 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<TestCase>
-    <DBMS>Any</DBMS>
-    <ControlDataInsertion>
-    </ControlDataInsertion>
-    <Query>
-        SELECT COLUMN_NAME AS ColumnName,
-        DATA_TYPE AS DataType
-        FROM information_schema.columns
-        WHERE table_schema = 'analytics'
-        AND table_name = 'asmt_StudentAssessmentObjectiveFact'
-        ORDER BY ORDINAL_POSITION ASC;
-    </Query>
-    <Result>
-        <ColumnName>StudentObjectiveAssessmentKey</ColumnName>
-        <DataType>nvarchar</DataType>
-    </Result>
-    <Result>
-        <ColumnName>AssessmentKey</ColumnName>
-        <DataType>nvarchar</DataType>
-    </Result>
-    <Result>
-        <ColumnName>ObjectiveAssessmentKey</ColumnName>
-        <DataType>nvarchar</DataType>
-    </Result>
-    <Result>
-        <ColumnName>Namespace</ColumnName>
-        <DataType>nvarchar</DataType>
-    </Result>
-    <Result>
-        <ColumnName>StudentSchoolKey</ColumnName>
-        <DataType>nvarchar</DataType>
-    </Result>
-    <Result>
-        <ColumnName>SchoolKey</ColumnName>
-        <DataType>varchar</DataType>
-    </Result>
-    <Result>
-        <ColumnName>StudentScore</ColumnName>
-        <DataType>nvarchar</DataType>
-    </Result>
-    <Result>
-        <ColumnName>ResultDataType</ColumnName>
-        <DataType>nvarchar</DataType>
-    </Result>
-    <Result>
-        <ColumnName>ReportingMethod</ColumnName>
-        <DataType>nvarchar</DataType>
-    </Result>
-    <Result>
-        <ColumnName>PerformanceResult</ColumnName>
-        <DataType>nvarchar</DataType>
-    </Result>
-</TestCase>
