--- src/EdFi.AnalyticsMiddleTier.Tests/TestCases/StudentAssessmentObjectiveFact/MP_2013_Mathematics_Seventh_Grade_605213_255901044_should_have_ResultDataType.xml
@@ -1,19 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<TestCase>
-  <DBMS>Any</DBMS>
-  <ControlDataInsertion>
-  </ControlDataInsertion>
-  <Query>
-      SELECT ResultDataType
-      FROM analytics.asmt_StudentAssessmentObjectiveFact
-      WHERE StudentObjectiveAssessmentKey='392-Seventh grade Mathematics-48-MP-2013-Mathematics-Seventh grade-zXFMqaQUouumDhHF8VnSmBJPaZ1Xbc8f/Li1nh1q-uri://ed-fi.org/Assessment/Assessment.xml'
-      AND StudentSchoolKey='605213-255901044'
-      AND SchoolKey='255901044'
-      AND StudentScore='1'
-      AND ResultDataType='Decimal'
-      AND ReportingMethod='Raw score';
-  </Query>
-  <Result>
-      <ResultDataType>Decimal</ResultDataType>
-  </Result>
-</TestCase>
